dse: Move manipulation routines to other module
